Unify GAS validators so they are part of the form definition
