+++ src/server/index.js
@@ -19,6 +19,10 @@ import csp from '~/src/plugins/content-security-policy.js'
 import sso from '~/src/plugins/sso.js'
 import { formsService } from '~/src/server/common/forms/services/form.js'
 import { outputService } from '~/src/server/common/forms/services/output.js'
+import {
+  formSubmissionService,
+  loadSubmissionSchemaValidators
+} from '~/src/server/common/forms/services/submission.js'
 import { catchAll } from '~/src/server/common/helpers/errors.js'
 import { requestLogger } from '~/src/server/common/helpers/logging/request-logger.js'
 import { setupProxy } from '~/src/server/common/helpers/proxy/setup-proxy.js'
@@ -34,8 +38,6 @@ import LandParcelPageController from '~/src/server/land-grants/parcel/land-parce
 import SubmissionPageController from '~/src/server/land-grants/submission/submission-page.controller.js'
 import { formatCurrency } from '../config/nunjucks/filters/format-currency.js'
 import { router } from './router.js'
-import { formSubmissionService } from './common/forms/services/submission.js'
-import { loadSubmissionSchemaValidators } from './common/forms/gas/gas-answers.schema.js'
 
 const SESSION_CACHE_NAME = 'session.cache.name'
